Setup modal for plan leave
